ci: use version as image tag when deploying if possible (#676)
This PR uses the image "version" (e.g. `1.2.0`) as the docker image tag, instead of the git commit hash. Also, the name `image_sha` was misleading since Docker images/layers also have a sha identifier.

My goal is to improve "readability" in order to make it easier to check which version is running in the cluster, while still having automatic CD of course.